feat(Logs): Add a reusable Logging interceptors.
Adds a logging interceptor annotation (@Logged).
Only works on non-private methods, on classes that are not final and
that have a non-private constructor.

Performance is about what you'd expect from any use of AOP method
interceptors (about 1/2 a ms per call) for methods that *are not*
logged. When the intercepted method is logged, the delay is closer to
a few milliseconds, up to even a tenth of a millisecond.

You can test the performance on your local server by running the
HelloWorld server and hitting the 'localhost:port/performance/{anything}'
endpoint.