defect(mongoId): MongoStyleIdGenerator cannot find mac address (#5)
The generator fails to obtain a MAC on macOS devices running
PulseSecure while PulseSecure is active (in VPN mode).

The network returned for the 'localhost' network interface is null.
Which is odd. But whatever. Now we search through the list until we
find a mac address. Generally it will be the same address on each
application restart.