Fix(Client): Response filter should not attempt to log large responses. (#8)
Response filter should not attempt to log large responses. Especially when logging is off.

Removed CapturingClientResponseFilter.java as it is a touch dangerous for a framework to support.
Filter is now a no-op if the Log Level will suppress the output.
Filter now limits response size to 2kb by default, but is configurable via a system property.